# projet-phase2/docs/api/brokerx-openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8085","description":"Generated server url"}],"paths":{"/api/v1/orders/{orderId}":{"put":{"tags":["order-controller"],"operationId":"replace","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplaceOrderDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderMutation"}}}}}},"delete":{"tags":["order-controller"],"operationId":"cancel","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CancelOrderDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderMutation"}}}}}}},"/api/v1/orders":{"post":{"tags":["order-controller"],"operationId":"place","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaceOrderDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Ack"}}}}}}},"/api/v1/deposits":{"post":{"tags":["deposit-controller"],"operationId":"deposit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepositDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["signup-controller"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1/auth/signup/confirm":{"post":{"tags":["signup-controller"],"operationId":"confirm","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ConfirmResult"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","parameters":[{"name":"X-Forwarded-For","in":"header","required":false,"schema":{"type":"string"}},{"name":"User-Agent","in":"header","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Response"}}}}}}},"/api/v1/orders/{orderId}/notifications":{"get":{"tags":["order-query-controller"],"operationId":"notifications","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NotificationDto"}}}}}}}},"/api/v1/orders/{orderId}/executions":{"get":{"tags":["order-query-controller"],"operationId":"executions","parameters":[{"name":"orderId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExecutionDto"}}}}}}}},"/health":{"get":{"tags":["health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/actuator/health":{"get":{"tags":["health-controller"],"operationId":"health_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}}},"components":{"schemas":{"ReplaceOrderDto":{"required":["expectedVersion"],"type":"object","properties":{"expectedVersion":{"type":"integer","format":"int32"},"qty":{"type":"integer","format":"int64"},"limitPrice":{"type":"number"},"duration":{"type":"string"}}},"OrderMutation":{"type":"object","properties":{"orderId":{"type":"string","format":"uuid"},"status":{"type":"string"},"version":{"type":"integer","format":"int32"},"timestamp":{"type":"string","format":"date-time"}}},"PlaceOrderDto":{"required":["accountId","clientOrderId","side","symbol","type"],"type":"object","properties":{"accountId":{"type":"string","format":"uuid"},"side":{"type":"string"},"type":{"type":"string"},"symbol":{"type":"string"},"qty":{"type":"integer","format":"int64"},"limitPrice":{"type":"number"},"clientOrderId":{"type":"string"}}},"Ack":{"type":"object","properties":{"orderId":{"type":"string","format":"uuid"},"status":{"type":"string"}}},"DepositDto":{"required":["accountId","amount","idempotencyKey"],"type":"object","properties":{"accountId":{"type":"string","format":"uuid"},"amount":{"minimum":0.01,"exclusiveMinimum":false,"type":"number"},"idempotencyKey":{"type":"string"}}},"Result":{"type":"object","properties":{"newBalance":{"type":"number"},"txId":{"type":"string","format":"uuid"}}},"SignupDto":{"required":["address","country","dateOfBirth","email","fullName","password","phone"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"fullName":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"country":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"}}},"ConfirmDto":{"required":["accountId","otp"],"type":"object","properties":{"accountId":{"pattern":"^[0-9a-fA-F-]{36}$","type":"string"},"otp":{"type":"string"}}},"ConfirmResult":{"type":"object","properties":{"accountId":{"type":"string","format":"uuid"},"status":{"type":"string"}}},"LoginDto":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"otp":{"type":"string"}}},"Response":{"type":"object","properties":{"sessionId":{"type":"string","format":"uuid"},"expiresAt":{"type":"string","format":"date-time"},"token":{"type":"string"}}},"NotificationDto":{"type":"object","properties":{"notificationId":{"type":"string","format":"uuid"},"type":{"type":"string"},"channel":{"type":"string"},"status":{"type":"string"},"payload":{"type":"string"}}},"ExecutionDto":{"type":"object","properties":{"executionId":{"type":"string","format":"uuid"},"orderId":{"type":"string","format":"uuid"},"counterOrderId":{"type":"string","format":"uuid"},"qty":{"type":"integer","format":"int64"},"price":{"type":"string"},"side":{"type":"string"}}},"CancelOrderDto":{"required":["expectedVersion"],"type":"object","properties":{"expectedVersion":{"type":"integer","format":"int32"}}}}}}